Implement Loader Functionality
Implemented the functionality to display loader. The loader variable is stored is
redux store and changes every time any ajax request is fired and completed. The loaderReducer
supports actions to show and hide the loader, both of these change the state of the loader in
Redux store. A seperate component has been created for loader which currently contains the boilterplate
code and checks if the loader variable is true or false.