Add movieStats table
Added movie stats table to show all the movies by actor. Implementation is achieved
by creating a movieStats component which contain its own state. The state is responsible for
keeping the track of index which changes to 10 at each render. The state is used to implement
the show more functionality.