Add Back Button
Added back button on actor page. The back button is added in search component. The search
component conditionally adds a class to input element which reverses the order of input element
putting the search icon on the right. Based on the same condition, the back icon is rendered in the
searchBar.