Add Count Badge, Fix text source for tab switches.
Add Count Badge

Added a count a badge in tabbed navigation under Actor Card component. The count
badge display the number of movies, tv shows, and awards for each actor.

Fix text Source for tab switches

The tab switches were programmed to changed the state of tab navigation using the
event.target.textCourse. This broke the code as I added the Count Badge. Because if you
clicked the count badge, the text source changed. Issue if fixed by providing a data-attribute
to every li element and that data-attribute is used as a source text for the state.